refactor: move editor state into editor route
The index page is used for display.
