+++ source-code/website/src/pages/editor/@host/@organization/@repository/index.page.tsx
@@ -8,10 +8,10 @@ import {
 	inlangConfig,
 	referenceResource,
 	repositoryIsCloned,
-} from "@src/pages/editor/state.js";
-import { Layout as EditorLayout } from "@src/pages/editor/Layout.jsx";
+} from "./state.js";
+import { Layout as EditorLayout } from "./Layout.jsx";
 import type * as ast from "@inlang/core/ast";
-import type { EditorRouteParams } from "@src/pages/editor/types.js";
+import type { EditorRouteParams } from "./types.js";
 
 export const Head: PageHead = (props) => {
 	const routeParams = props.pageContext.routeParams as EditorRouteParams;
